refactor : toString 사용 없이 printMovement 하도록 수정
