+++ src/main/java/view/OutputView.java
@@ -12,8 +12,7 @@ public class OutputView {
         System.out.println(String.join(", ", winners) + "가 최종 우승했습니다.");
     }
 
-    public void printCarsMovement(List<String> movement) {
-        System.out.println(String.join("\n", movement));
-        System.out.println();
+    public void printMovement(String name, int forwardCount) {
+        System.out.println(name + " : " + "-".repeat(forwardCount));
     }
 }
