+++ src/test/java/domain/JudgeTest.java
@@ -1,4 +1,4 @@
-package service;
+package domain;
 
 import domain.Cars;
 import domain.Judge;
